chore: Extract a HumanAuthentications controller
